transport: split decoder into own module
